requirements.txt
boto3==1.16.34
botocore==1.19.34
Brotli==1.0.9
click==7.1.2
Flask==1.1.2
Flask-Compress==1.8.0
Flask-SQLAlchemy==2.4.4
future==0.18.2
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mysql==0.0.2
mysql-connector-python==8.0.22
mysqlclient==2.0.2
plotly==4.14.1
protobuf==3.14.0
python-dateutil==2.8.1
python-dotenv==0.15.0
retrying==1.3.3
s3transfer==0.3.3
schedule==0.6.0
six==1.15.0
SQLAlchemy==1.3.21
urllib3==1.26.2
Werkzeug==1.0.1
wget==3.2
